Remove the namespace global
We need to be able to switch into other namespaces
temporarily in order to scan the body of functions/methods
in their native namespaces and switch back when we are done.
By passing the namespace around and returning the new namespace
from the stmt_list passes we can do that in a much cleaner way.