Assign types from is_a series of functions
The is_a (is_string, is_array, etc.) series of functions guarantees,
when used in an if block, that the type of the variable matches the type
checked for. Add a simplistic check that works for the simple usage of
checking only the type. Only works for simple function call, not
expressions that include the is_a check.