Adding stricter return type documentation checking
Phan now ensures that the type specified in a docblock matches the type
in a PHP7+ return type hint. For example:

/**
 * @return string
 */
function fn() : int {
    return 7;
}

will produce a PhanTypeMismatchReturn issue, as the docblock return is
of type string and the hint is of type int.

Phan makes an exception for returning a more narrowly-typed array, e.g.

/**
 * @return SomeType[]
 */
function fn() : array {
    return array(new SomeType(), new SomeType());
}

will be considered as returning SomeType[].