Handle aliased namespaces correctly
Aliased namespaces were not being handled correctly, per issue #329.
This was broken in 1121ab9. This seems to have snuck in due to an
incorrect test. The expectation for 0136_alias_with_suffix.php claimed
to have no issues, but running the file through php results in:

  PHP Fatal error:  Uncaught Error: Class 'NS2\C\D' not found in ...

Patch adjusts the test to be correct, and makes the necessary update in
Type.php to pass through the namespace prefix so that the alias is
properly resolved.

Fixes #329