Fix bugs in type checking Generators, stricten checks
Parse the full source of the function/method/closure
**before** analyzing any return statements.
This ensures that phan will properly report buggy functions
that have an incorrect return type of `Generator` (or Traversable, or
Iterable).

If the function-like is a Generator, type check against declared type
exactly once.

It's also possible to have a `return` statement before `yield`
(`if($cond) {return; } yield;`), which this didn't handle.

The old code also failed to analyze the `yield from` case (`visitYieldFrom`)

Aside: It doesn't seem like phan will have Generator types in the near future.
https://github.com/phpDocumentor/fig-standards issue 5 discusses that
proposal.

Possibly related to #355 and #399