Replace nightly CI build with 7.1RC6 CI build
It's a bit silly to be building phan against PHP nightly since we depend so heavily on the php-ast extension.

Instead, let's build against 7.1RC6. PHP 7.1 will be the next version of PHP that we will need to support in Phan and it's a good idea to start tracking the incompatibilities now. The 7.1RC6 build will allow failures just like the nightly build did.