Don't issue PhanTypeReturnMissing if the method always throws
This isn't a complete implementation of the 'only possible to throw'
definition, but seems like a reasonable first start. If the only way
to exit a function is a throw then the return isn't missing, and it
will never return the wrong value.

This is usefull for patterns where a base class defines a method
with a return type, but the implementation is to throw a
NotImplemented exception.