Detect undeclared constants/class constants in array literals.
This was seen in a large array used for string lookup.

As well as values past the first 5 (Only types of the first 5 array values are
analyzed by phan)

Aside: function calls in array literals are also not analyzed. This
seems less important.

This will possibly emit duplicate exceptions (UnionTypeVisitor visits
most uses of class constants, but not all.

Add tests related to edge cases of class constants and constants