Create scoped variables from if(isset($..))
Inside an if(isset($foo)) block the variable $foo should exist. It looks
like the code was already there, just commented out. The only change i
made was to allow a new variable to be created if the scope didn't know
anything about it.

There is an edge case that doesn't yet work though, the context doesn't
seem to be appropriately poped and the variable detected as set by the
if is believed to continue existing further. Not sure how to handle
allowing assignment in an if statement to continue, but not this.