Account for backslashes in Phan::isExcludedAnalysisFile
Fixes #489

I have gone for a simple approach doing a str_replace
on all paths that pass through the isExcludedAnalysisFile
method.

Other posibilities would be to check if the path contained
a \ and then do the replace, or check the OS that the code
is running on and only do the replace on Windows.