Smoother progress bar
Call fwrite() once - hopefully less likely that hitting enter will split
this up in the middle of a progress bar, terminal should be refreshed
less often, etc.

In preparation for the daemon mode, cache the results of
shouldShowProgress() so that config instance isn't fetched multiple
times per instance in the method phase.