Add magic property annotations to phan
Currently, @property, @property-read, and @property-write are all
equivalent, due a to lack of checks if a context node is getting/setting a
property.

Also add annotation @phan-forbid-undeclared-magic-properties
(Motivation: if a class is expected to have complete annotations for its
getters/setters, then it is useful to warn if something undeclared is used
in that class, to prevent typos and to ensure analysis maintains a
consistent quality)

Limitations:

- Does not warn if __get/__set don't exist.
- Not able to have a __get() shadow a property of the same name inside
  the current class (e.g. getter of private property, which does
  initialization)

Fixes #386

Taken from my fork, with the new @PhanForbidUndeclaredMagicProperties
annotation introduced.

add tests of property-read and property-write's expected behavior
(currently implemented the same way as property)