Fix casting types to/from nullable - use static::, not self::
For #567 and #582

self::make() returns an instance of the class Type (or a generic array),
but we instead need the same subclass as the original.
static::make() returns an instance of **the same** subclass of Type, e.g.
ArrayType.

ArrayType::withIsNullable(false) was returning a non-array

Also, add a small optimization and check if is_nullable would be
the same as before (e.g. for plugins)

Also, add tests.