Finish implementing @phan-forbid-undeclared-magic-methods
Fixes issue #467

This was parsed, but code changes (In places which allowed undeclared
methods if __call/__callStatic existed) weren't yet made.

Also, add a test for warning about static call to non static methods.

This PR doesn't bother checking if __call and __callStatic are declared
with any of these annotations.
The implementation assumes that they those methods have been declared
in the class (Or in subclasses of the class with the required magic methods),
and that the annotations are kept up to date.

- This is meant to make it as easy as possible to start using phan with
  frameworks such as phpunit mock objects, or laravel, etc,
  where interfaces or abstract classes declare magic methods.