Rudimentary coloring support for the default (plain text) printer.
To use it, invoke phan with `--color`.
These colors are arbitrary, I couldn't think of a well known color scheme to
imitate for this.

For Issue #160.
The message would be colorized for other printers, which is usually
undesirable, but not bothering to check that (e.g. in xml).

Choose color based on issue severity (Red for severe, pink/orange for normal,
yellow for minor)

- Bold/italic with the same color might make more sense

Further work

- color based on regex for constant names, etc.
  (E.g. multiple colors for the class and constant of `\A::constname`)
- 32-bit color, if supported by the terminal
- Allow multiple directives, e.g. to set the background color.
- (This assumes a dark terminal background.
  Adding --dark-colorscheme in the future will be useful)