Fix for uncaught exception converting generic array type to/from nullable.
Found the cause: The generic array is converted to nullable when
checking if the parameter default already has a nullable version
in the parameter type.

For #665

The issue is that new static(namespace, name, ...) in Type::make (Called
because of withIsNullable()) would end up invoking
GenericArrayType::__construct(Type $inner), throwing a PHP Error

There are some issues to solve, e.g. whether or not the uncommon
?string[] is a meaningful annotation. This is the easiest/fastest
way to stop throwing an Error.

Also, that seems to be ambiguous, since it could be interpreted
as (?string)[] or ?(string[]), and converting the type to/from
string would suffer from that ambiguity.
(Currently, Phan parses the array first, then the inner nullable type)