Improve PhanParam*RealMismatch error messages, fix bugs.
Include a cause for the incompatibility, to make fixing the
incompatibility easier.

Methods returning reference and not returning references can be
mixed in overrides, without issue.

getNumberOfOptionalParameters() can be influenced by the presence of
func_get_args() and variadic params in the overriding method (and
overridden). Move away from that.

getNumberOfRequiredParameters() can be influenced by out of date method
signatures in phan's function type info.
Also, if a module is missing reflection info and has classes that can be
extended, it is permitted to have 0 required parameters in a method override.