Emit issues when checking for common typos in phpdoc annotations
Fixes issue #778

`@returns` is sometimes accidentally used by people familiar with JSDoc.
Also note if an annotation occurs in a place it isn't expected to.
(In the future, this may be configurable)

Fix various issues

Note: May want to do this differently if other analyzers are used
(e.g. if they use `<something:something>` in the types