Allow checking if phpdoc param types are narrower than corresponding signature's types
Emit PhanTypeMismatchDeclaredParam if any of the types in the phpdoc
union type are incompatible or wider (compared to the actual signature).

Also introduced in this PR: if phan is configured with
`prefer_narrowed_phpdoc_param_types`, then prefer phpdoc types if
they're compatible and narrower with real types.
(E.g. analyze methods as if they had the signature `Type[]` over `array`)

Fix issues this new config setting detected within Phan itself.

Related to #505 , but for params