Fixes #871: Improve warnings and inferences about function references
Fixes #958 (already "fixed" on master branch, but had different issues)

The internal function signature map prefixes references with "&w_" to
indicate that the a reference is write-only
(Previous value is ignored, and overridden unless there was an error)
It also adds "&rw_" to indicate that the original type should be preserved
without adding new types (e.g. for sorting methods)

- No way to specify this for user-defined functions/methods, yet.

This may change even more before the 0.9.3 release,
but the preg_match behavior and sort() tests should ideally continue passing.