Fixes #926 : Analyze @override in phpdoc of methods and class constants
Create `CommentOverrideOnNonOverrideMethod` and
`CommentOverrideOnNonOverrideConstant`.
These issue types will be emitted if `@override` is part of doc comment
of a method or class constant which doesn't override or implement anything.

(Additionally, this won't warn about magic methods,
such as __construct(), __get() or __set(),)

Annotate demo plugins with @override

This makes it obvious to plugin developers that the examples are
overriding methods.

Phan will emit an error if those methods are misspelled or if the base
method are removed (unlikely).