Fixes various bugs in the implementation of `@phan-closure-scope`
- Add `@phan-closure-scope` to be consistent with names of other custom
  annotations. Leave in `@PhanClosureScope`, which will be removed eventually.

Fixes #309 : `@phan-closure-scope` should be stable enough to be used to analyze
code using Closure::bind and Closure::bindTo

- The place where the closure is used may be in a completely different
  file/module from the place where the closure is declared.

Fixes #590 : Emit the correct file name for closures that are annotated
with `@phan-closure-scope`

Fixes #790 : Didn't import variables with `use` when `@phan-closure-scope` was
added to a function.

---

This fixes issues in the original approach I created earlier for #309
Overriding the **entire** Context meant that the file name for issues was wrong, and
the variables were unavailable.
Overriding only the class FQSEN is closer to the intended behavior