Fixes #978: Add `\Closure` to the type system.
Add it as a first class type, and split it out of `callable`.
Allow `\Closure` to cast to `\Closure` and `callable`, in addition to
`object`. Don't allow `callable` to cast to `\Closure` without checking
if it's an instance of `\Closure` (add test).

Unrelated: Make any subclasses of Type final if they currently have no subclasses.

Update 2 tests expectations to expect the name `\Closure` instead of
`callable`. Analysis of parameter passed to closures continues to work.