Fixes #1027 : Improve block exit status checking
Allow 2^5 - 1 (31) combinations of possible exit statuses for a given
statement or statement list.
(0 is an invalid combination of bits. Even inferred infinite loops and
termination have a non-zero status (currently "return").

A statement can be inferred to have any combination of
proceed/continue/break/throw/return.

- If a statement does not have "proceed", then the statement immediately
  after this statement (if any) is unreachable.
- If a statement has only "proceed", it is assumed to always proceed.

Arbitrary function calls, other fragments of code are assumed not to
throw, except for the `throw` statement.