Fix for type narrowing bleed for ternary conditional in global scope
Fixes regressions in solution to #538 for ternary operators.

This bug would lead to erroneously inferring the types from
the right hand side of a ternary operator in the global scope,
in statements following a ternary operator.

(Cloning the global scope has no effect;
it has static variables, not instance variables)

Add tests.