Support indicating that a reference parameter's input value is unused
This can be done in PHP code by writing `@phan-output-reference`
on the same line as an `@param` annotation.
This indicates that Phan should not warn about the passed in type, and
should not preserve the passed in type after the call to the
function/method.

- (In other words, Phan will analyze a user-defined reference
  parameter the same way as it would `$matches` in
  `preg_match($pattern, $string, $matches)`)

Example usage:

```php
/** @param string $x @phan-output-reference */
function set_x(&$x) { $x = 'result'; }
```