Improved generic array types inferences of array literals.
Get rid of the 5 element limit.
Analyze the union types of all array elements instead of just the first 5.
Add `mixed` to the types of keys/values when a single element can't be
inferred.

This is being done in preparation for supporting array shapes such as
`array{fieldName:string}`.
(The alternative would be to have the same/different limitation on array
shape sizes.)

This slows down Phan slightly.

Fixes #985