Fixes #1408: Improve analysis of try/catch/finally blocks
Analyze catch blocks with the inferences about the try block.
(Phan pretends the entire try block succeeded or failed and uses the
combination of possible union types for analysis)

Analyze finally blocks with the inferences about the try block
(and the following catch blocks).
When there is a finally block and at least one catch block,
Phan merges the inferences about the entire try block succeeding
with inferences about subsequent catch blocks.

(And some more heuristics)

When `declare(strict_mode=1)` is enabled,
Phan requires that variables be defined in all possible blocks to work
properly.

Rewrite code with a potential bug (false positive) to be clearer.
(in PreOrderAnalysisVisitor)