Add another AST simplifier rule
Get rid of the rules which are no longer needed.
BlockAnalysisVisitor has the same awareness of a block of statements
failing to return (e.g. in an `if` branch) that ASTSimplifier would,
and accounts for that when choosing which inferences to merge.

Fix a bug which was making ASTRewriterTest fail to do anything
(Wrong file name used for original input file)