Fix error with class hydration; add --language-server-analyze-only-on-save
Previously, there would be an uncaught CodeBaseException due to phan
failing to resolve a missing/inherited constant in the parse phase.

This defers hydration until after Phan finishes parsing a given class.
(This isn't ideal, and won't work as expected in some edge cases
when using inherited constants)

Fixes #1325: Allow language server to only be notified on file save
(and not on file edit). This greatly reduces CPU usage.