Fixes #819: Make language server work on Windows
The `--language-server-allow-missing-pcntl` option must be provided by
the language client for this to work.

This implements a fallback mode without the dependency on pcntl

- Originally, pcntl_fork was used to avoid the need to manually
  save and restore state of Phan's data structures for
  classes/properties/constants/files/issues.

  (Phan's codebase may change or add new properties)

- This PR makes Phan manually save and restore it's representation
  of the project,
  only when this CLI setting is used an pcntl is not available

  This is meant to be limited to the mutable parts.

Make FQSENs throw if we attempt to serialize/unserialize/clone them.
They have to be globally unique.

Fix a few bugs in the previously unused codebase cloning methods.

Fix an edge case in the tolerant php parser to php-ast converter