Implement support for "go to definition" for class names.
For #1483

Modify TolerantASTConverter to be subclassable.
Add a subclass which will track which node is
having the cursor searched for.

This PR implements support in the language server support, but not in
"daemon mode". It's possible to add that later.

Add a passing integration test of "go to definition".
(i.e. `textDocument/definition`)