Improve analysis of `@template` on functions&methods
Allow using `@template` on functions, closures, and instance/static
methods. Previously, this was only allowed on classes.

Support `@param T[] $x` as a way to indicate a function expects an array of
the template parameters.

- Support for extracting templates from other types
  (e.g. `MyClass<T>`, `iterable<T>`, `callable():T`, `array{0:T}`, etc.
  is not implemented yet)

Be more strict about `@template` on class. Warn if the constructor doesn't
reference the template in PHPDoc (instead of just checking if positions
were used)

Related to #522