Add `@phan-assert` and the plugin PHPUnitAssertionPlugin
This is an early design. The API will probably change,
and an easier way to use Phan assertions
from plugins will probably be added in the future.

The behaviors of these assertions may change in the future.
(e.g. narrow some types instead of setting to those exact types)

More types of assertions may be added in the future.

- `@phan-assert int $x` will assert that the argument to the parameter `$x` is of type `int`.
- `@phan-assert !false $x` will assert that the argument to the parameter `$x`
  is not false.
- `@phan-assert !\Traversable $x` will assert that the argument to
  the parameter `$x` is not Traversable (or a subclass/subtype of that)
- `@phan-assert-true-condition $x` will make Phan infer that the argument
  to parameter `$x` is truthy if the function returned successfully.
- `@phan-assert-false-condition $x` will make Phan infer that the argument
  to parameter `$x` is falsey if the function returned successfully.
- This can be used in combination with Phan's template support.

Add `PHPUnitAssertionPlugin` to make Phan aware of the side effects of a
limited number of helper methods that PHPUnit exposes in test cases.