Make variable and static property completions compatible with VS Code
This requires passing in the flag `--language-server-completion-vscode`

- Without this flag, variable/static property completions
  work in the client LanguageClient-neovim,
  but not with VSCode (for variables/static properties)
- The new code works in the client vscode-php-phan, but not in neovim
  (for variables/static properties)

I assume this is because of differences in the ways they identify the
current token being completed (vim doesn't include the `$`, VSCode does)

- A short term solution is to add client-specific workarounds instead

TODO: Look into implementing textEdit;
hopefully client support is better with textEdit in all platforms.
Unfortunately, we don't currently keep the column information,
which that requires.