Avoid repeatedly calculating the terminal's width
If the COLUMNS environment variable isn't explicitly set (most people probably
don't), then Symfony Console will shell out to stty and do a grep, and the
proc_open() will cause a fork of a potentially 2+ GB process. If the instance
is short on memory, the kernel will reject the fork since the child will not
be able to allocate all that memory.

Instead of constantly calculating the width of the terminal each time the
progress needs to be updated, just calculate it once and save that for the
rest of the process. It's possible that people might resize their terminal
window while phan is running, but that is a narrow edge-case.

Credit to @hashar who identified the performance bottleneck at
<https://phabricator.wikimedia.org/T219114#5084302>.