Add PassByReferenceVariable::getElement()
This public method can be used to read the $element property of the class. I understand that this class is meant to just mimick it, but there may be cases where we need the full element object. For instance, I'd need it in the SecurityCheckPlugin. Yes, the plugin is a bit hacky, but I guess a public getter can't cause any harm.

[0]: https://github.com/wikimedia/phan-taint-check-plugin