Increase max recursion depth to 3
And put it inside a class constant.
I checked the runtime on the whole MediaWiki core, both with max = 2 and max = 3. In both cases, I ran phan twice, and the times were exactly the same: 1 min 15 sec. If there's no performance loss on a project this big, I believe there's nothing to be afraid of.