Demo plugin: handles nonexistent classes in if(class_exists()) blocks
Usecase: some software uses an optional third-party library
(for example, a code highlighting tool) and can work without it,
so it uses the following approach:

if (class_exists("HighlightCode")) {
    echo HighlightCode::apply($text);
} else {
    echo $text;
    $logger->debug("No code highlighting (library is not installed)");
}

A example of such software is MediaWiki extension PageForms, which
unlocks extra features if "Semantic MediaWiki" framework is installed,
but can work without it,
because these features are within if(class_exists()) blocks.

Currently Phan would always emit an error if the class doesn't exist,
meaning PageForms can only be checked by Phan when "Semantic MediaWiki"
is installed, and can't be checked without it.

This plugin tracks "if" blocks that are conditional on
class_exists("ConstantString"), and if "ConstantString" class is
missing when Phan is invoked, then suppress the Phan issues related to
missing methods within the ConstantString class, but only
if those methods are used WITHIN such an if(class_exists()) block.

Note: "missing method" issues are NOT ignored if "ConstantString" class
exists at the time when Phan is invoked.

Possible future improvements:
1) detection of else blocks after if/elseif(!class_exists()),
they may/should be handled the same way as if(class_exists()).
2) Checks like if(method_exists()) may/should be handled the same way,
and the format of WithinSpecialConditionPlugin::$withinStack
(function name + array of arguments) was chosen with this in mind.