Fix false positives that are emitted when --no-progress-bar is used
The memoized methods of Union Types got cleared when Phan started a
different analysis phase, to deal with false positives when
inheritance wasn't analyzed yet.
However, CLI::progress would not get called for some phases when there
was no progress bar, so this didn't work all of the time.

Fixes #3677