Bugfix: Properly recursively analyze inherited methods
Fixes #3740

Phan was modifying the parameter list and scope variables of the cloned
method in the subclass, but then was analyzing the original method in
BlockAnalysisVisitor instead.

Instead, modify the parameter types of the original definition before
analyzing the method, all of the time.

Show empty union types as `(empty union type)` instead of an empty
string, in most issue messages.