Also check provided path against project realpath
This affects the language server on Windows
when the drive letter is lowercase.

This may also affect plugins for phan that use getProjectRelativePath
for the real path.

Related to https://github.com/phan/phan/pull/4136 .
Some projects may be symlinks, AND rely on `../../` in their
.phan/config.php the way users have set up the project,
so try to continue to accommodate that.
(Alternately, manually normalize "./", "//", etc. without resolving
symlinks if the input wasn't an absolute path.

TODO: Or figure out how to correctly normalize non-symlinks in
case-insensitive filesystems, or even just the drive letter
from realpath().