Be stricter when checking if `__construct` exists
Noticed when investigating #4320 that InferPureVisitor
would cause the placeholder method to be created (depending on config
settings), which would cause the warning about a missing constructor to
fail to be emitted, even with no conditionals.

The original bug report had `is_callable`, which Phan currently does not
use to infer the existence of methods. See #4186