[WIP](broken) intersection type parsing support
For #1629

Previously, intersection types were parsed for compatibility, but
`A&B` was treated as an alias for `A|B` because there was no support in
the type system for intersection types

There is a LOT more work remaining to actually support the type casting
logic for intersection types.

The union type parsing is done with a regular expression, and it may be
better to replace it with a real parser at some point.

As a workaround, support `and<T1, T2, T3>` in phpdoc for union types and
translate to that.
(TODO: Change to `phan-intersection-type<T1, T2, T3>` if intersection
types work well enough in phpdoc)

TODOs:

- Update everything that checks for types that correspond to a single fqsen
  to also check for intersection types
- Support getting method signatures from intersection types
  (to start, just look for the first method that implements it)
- Eliminate redundant types from intersection types
  (E.g. ArrayObject&Countable -> ArrayObject)
- Handle self/static (resolve in context, etc)
- Throw an IssueException for non-interfaces or non-classes
- Infer intersection types from conditions narrowing interface/class
  types and fix test failures or update expected messages
- Fix test failures
- Add more tests
- Implement actual checks to allow casting A&B&C to A&B
- Allow casting ArrayObject to ArrayAccess|Countable
  (i.e. if it inherits all types)

  This may or may not require refactoring
- Override various methods implemented on atomic types
- For compatibility, add a config setting to temporarily turn this off
  for projects that rely on the old behavior or that encounter issues
  with the intersection type logic.

Improve accuracy of type casting rules

Continue working on intersection type support

WIP check for intersection types in more places

Fix warnings, update test expectations, add TODOs

WIP(broken) require codebase for type casting checks