ConfigPluginSet: don't nest each variable info closure
The old code would, for each plugin that wanted to register a mergeVariableInfoClosure, replace the $mergeVariableInfoClosure with a new function that called the plugin's closure, as well as the old $mergeVariableInfoClosure. This meant that a new level of nesting was introduced for each plugin registering such a closure after the first.
Instead, use a single closure for all plugins.