Make line numbers in polyfill consistent with native ast
E.g. for

	return [
		'foo' => ['bar']
	];

It seems like the way php (and as a result php-ast) is getting the line number
for anything that has 1 or more children is to use the line number
of the first non-null child. So the first child of the AST_RETURN
is the AST_ARRAY, and the first child of the AST_ARRAY is the AST_ARRAY_ENTRY,
and the first child of the AST_ARRAY_ENTRY is the AST_ZVAL `'foo'`
(which is just a regular value in php-ast's representation,
the AST_ZVAL kind is not included as a exposed constant),
so php itself would infer line 5 for the return statement
(and any of its error messages).

- If there are none of those, php uses whatever the current line number of the
  lexer would be.

I was hoping php would eventually improve on that, but that hasn't
happened yet.

Closes #4520