Don't hydrate classes until all aliases are resolved
When analysing using `enable_class_alias_support`, it is possible to
have two aliases which refer to each other:
```
class_alias(A::class, 'OldA');
class B extends OldA { ... }
class C extends B { ... }
class_alias(C::class, 'OldC');
```
If phan processes the OldC => C alias first, it would hydrate C and
then B in the process.  But when hydrating B, the OldA => A alias isn't
loaded yet, so any properties or methods defined in A won't be loaded
into B or C.

The solution is simply to defer hydrating classes until after all of
the aliases have been processed.  This ensures that when C and B are
eventually hydrated, the 'OldA => A' alias is known.

This was https://phabricator.wikimedia.org/T389698 in the MediaWiki
issue tracker.