PHPDocRedundantPlugin: detect redundant parameter lists
For now, this uses a stricter set of checks where the parameter list is
checked as a whole (i.e., all parameters must be redundant), and
excluding doc comments with unexpected features (like parameters after
return, or extra annotations between parameters).

Refactor the Fixer code so that it tries to preserve empty lines.