CallableParamPlugin: Do not report invalid callables for functions taking callable|foo unions
And likewise for class-strings for functions taking class-string|foo.

Fixes #1648.

Most of the complexity here is needed for union types that include
both class-string and callable, because we need to avoid emitting
issues about a value being incorrect for one type when it's correct
for the other, but still emit issues when a value is incorrect for
both types.