Fix template parameter type leaking out
Fixes #4982.

One common operation with generic classes is, given a union type,
computing a map of template params to concrete types used in this
type, so that this map can be applied to another union type.

A problem occurs when the given variable's type is a generic type
that is missing some or all of the type parameters (e.g. determined
from a native PHP type annotation without a PHPDoc comment):
let's say an instance of `X` which is really `X<T>`. The computation
of the map would duly report that 'T' parameter is the `T` template
type, this would be applied elsewhere, but it was useless, as `T`
doesn't mean anything outside of the scope of the `X` class.

What should happen in this situation, in almost all cases, is that 'T'
should be substituted to indicate that we know nothing about the type.

I'm making this change in the first step of this operation (computing
the map, or `getTemplateParameterTypeMap`) rather than the second
(applying the map, `withTemplateParameterTypeMap`), because it turns
out that the second one should, in almost all cases, not change types
that aren't listed in the map; for example, we often want to map a
class's template params in an inherited method, but leave the method's
own template params unchanged.