Clazz: Remove some dead code dealing with generics
I think this was intended to handle the case where a Generic2<X, Y>
extends a Generic1<A, B>, and the template parameter names in one
class definition have to be mapped to the names in the other one,
so that inherited methods and properties type-check correctly.
However, the code actually does nothing, it just maps X and Y to
themselves. The actual resolution of inherited template parameter
names happens in addMethod() and addProperty().