Add tests for an Option/Some/None type
Phan flags most of the mistakes with generic type parameters when
passing them to or returning them from functions, but it doesn't flag
the incorrect implementation of None as Option<null> rather than
Option<never>. It also fails to flag any of the mistakes with generic
type parameters when assigning to properties. No wonder our own `None`
that inherited from `Option<null>` has not been noticed for so long.