Map inherited template types in some more edge cases
Sometimes, the template types of properties and method parameters
that we've so carefully mapped would be overwritten later.

Their types are based on their PHPDoc and default value types, and
since the default values can depend on constants from classes that may
not be loaded yet when the code is parsed, the processing of *both*
PHPDoc and default values is deferred. (Property::getFutureUnionType(),
FunctionTrait::addParamsToScopeOfFunctionOrMethod())

In order to support the case where it happens after the class
inheritance, map inherited template types in the internal
representation of PHPDoc comments too.

This is ugly, I wish there was a better way to avoid this problem.