Rewrite asExpandedTypes[PreservingTemplate]() functions on all Types
Each Type subclass duplicated the code for asExpandedTypes() and
asExpandedTypesPreservingTemplate(), with occasional divergences that
caused bugs. They usually only differed by 2-5 lines (calling either
asExpandedTypes() or asExpandedTypesPreservingTemplate() recursively,
and either adding a template-erased type to the result or not).
Combine them into one function that takes a boolean parameter.

The implementation on Type itself had bugs, resulting in bogus parent
types being added when expanding generic types (because both a generic
and a non-generic parent type was present in getAdditionalTypes()).
Resolve this by removing parent types from additional types entirely;
parent types are already handled separately, so that was just
duplicating work. This is what causes the changes in tests - the bogus
types were printed in error messages and allowed some casts that
should have emitted issues. The handling for class aliases was also
impossible to understand and probably buggy, redo it.

Implementations on GenericArrayType and others should have made use of
the base implementation on Type to expand element types, but instead
they reimplemented it poorly. Notably, handling for parent types vs
additional types was different everywhere.

Centralize the boring memoization and recursion limits in the Type
class, and keep only the differing parts in each subclass.
Centralize the documentation there too, and actually document a
little bit about what these functions do.

Remove public methods UnionType::replaceWithTemplateTypes() and
Clazz::resolveParentTemplateType(), which were only used by the
removed code. I'm not sure what Phan's compatibility policy for
plugins is; is this okay to do, or should I only deprecate them?

I apologize for the huge commit, but this seems difficult to solve in
a more piecemeal fashion because of the three-way dependency between
Type::asExpandedTypes(), GenericArrayType::asExpandedTypes() and
Clazz::getAdditionalTypes(), all of which only worked somewhat right
due to bugs in the others.