Keep Method::hasTemplateType() more in sync with the params
There were some cases where we would replace all template types with
concrete types, but the flag on the method was not updated. This was
mostly harmless, but possibly inefficient.

There might have also been cases where a method was created with
template types in a more complex manner (e.g. nested deeper inside a
type parameter, or only in the unresolved PHPDoc type), but we don't
have a test case for that.